Convert hab_install to a custom resource
Closes #13
